+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/chat/GroupChatImpl.kt
@@ -8,6 +8,7 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.message.RawMessage
 data class GroupChatImpl(
     override val id: ChatId,
     override val title: String? = null,
+    override val description: String? = null,
     override val inviteLink: String? = null,
     override val chatPhoto: ChatPhoto? = null,
     override val pinnedMessage: RawMessage? = null,