+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/chat/abstracts/GroupChat.kt
@@ -2,6 +2,6 @@ package com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.chat.ChatPermissions
 
-interface GroupChat : PublicChat {
+interface GroupChat : PublicChat, DescriptionChat {
     val permissions: ChatPermissions?
 }